refactor: Avoid needing a behaviour constructor function
Previously, the user had to wrap the behaviour in a constructor function
to create a boxed behaviour. This is logically redundant as the
constructor function that should be used is already defined when the
user implements the specific type of behaviour for their struct. This
redundancy is now lifted through the use of an automatically implemented
`IntoBehaviour` trait.